Remove findDOMNode usage from the NavigableToolbar component
