+++ packages/block-library/package.json
@@ -45,7 +45,7 @@
 	},
 	"devDependencies": {
 		"deep-freeze": "^0.0.1",
-		"enzyme": "^3.3.0",
+		"enzyme": "^3.7.0",
 		"react-test-renderer": "^16.4.1"
 	},
 	"publishConfig": {
