+++ packages/components/package.json
@@ -49,7 +49,7 @@
 	},
 	"devDependencies": {
 		"@wordpress/token-list": "file:../token-list",
-		"enzyme": "^3.3.0",
+		"enzyme": "^3.7.0",
 		"react-test-renderer": "^16.4.1"
 	},
 	"publishConfig": {
