+++ packages/compose/package.json
@@ -28,7 +28,7 @@
 		"lodash": "^4.17.10"
 	},
 	"devDependencies": {
-		"enzyme": "^3.3.0",
+		"enzyme": "^3.7.0",
 		"react-dom": "^16.4.1",
 		"react-test-renderer": "^16.4.1"
 	},
