+++ packages/edit-post/package.json
@@ -45,7 +45,7 @@
 	},
 	"devDependencies": {
 		"deep-freeze": "^0.0.1",
-		"enzyme": "^3.3.0"
+		"enzyme": "^3.7.0"
 	},
 	"publishConfig": {
 		"access": "public"
