+++ packages/editor/package.json
@@ -61,7 +61,7 @@
 	},
 	"devDependencies": {
 		"deep-freeze": "^0.0.1",
-		"enzyme": "^3.3.0",
+		"enzyme": "^3.7.0",
 		"react-dom": "^16.4.1",
 		"react-test-renderer": "^16.4.1"
 	},
